Move scheduling code into scheduler module and use Mutex
