+++ src/arch/x86_64/interrupts/irq.rs
@@ -3,7 +3,7 @@ use device::keyboard::ps2::PS2_KEYBOARD;
 use device::pic_8259 as pic;
 use device::pit::PIT_TICKS;
 use device::uart_16550 as serial;
-use task::{Scheduling, SCHEDULER};
+use task::scheduler::{Scheduling, SCHEDULER};
 use x86_64::structures::idt::ExceptionStackFrame;
 
 pub extern "x86-interrupt" fn timer(_stack_frame: &mut ExceptionStackFrame) {
