+++ src/syscall/process.rs
@@ -1,5 +1,6 @@
 use alloc::String;
-use task::{ProcessId, Scheduling, SCHEDULER};
+use task::ProcessId;
+use task::scheduler::{Scheduling, SCHEDULER};
 
 /// Wrapper around scheduler.create() and ready() that can be called in processes
 pub fn create(new_proc: extern "C" fn(), name: String) -> ProcessId {
